refactor(core): host test_support in julie-core behind feature, break dev-dep cycle (ADR-0006)
Move test helper bodies into julie_core::test_support, gated by
#[cfg(any(test, feature = "test-support"))]. julie-test-support becomes a
thin re-export (pub use julie_core::test_support::*). julie-core stops
depending on julie-test-support — cycle broken, two-rlib mismatch gone.

Changes:
- Add crates/julie-core/src/test_support/{mod,db/mod,db/rows,db/tests,
  tempdir,cleanup}.rs — single source for all shared test helpers
- julie-core Cargo.toml: [features] test-support=["dep:tempfile"], optional
  tempfile dep; remove julie-test-support dev-dep (cycle break)
- julie-test-support: Cargo.toml reduced to one dep (julie-core +
  test-support feature); src/lib.rs is now 4 lines of pub use
- Delete crates/julie-core/src/tests/helpers.rs (local duplicate superseded)
- Repoint database.rs, database_row_mapping.rs, vector_storage.rs,
  database/concurrency_wal.rs from super::helpers::* + julie_test_support::*
  → crate::test_support::*

Gates:
- cargo build: green (full workspace)
- cargo build -p julie-core: green; tempfile absent from --no-dev-dependencies tree
- cargo nextest run -p julie-core: 118/118 pass (was 107; +11 builder/tempdir tests)
- cargo tree -p julie-core | grep julie-test-support: EMPTY (cycle gone)
- cargo tree -p julie-test-support | grep julie-core: shows one-way dep
- grep "pub fn file_info_builder": single source in test_support/db/rows.rs
- top-crate editing suite (103/103): re-export chain works
