+++ .memories/2026-06-03/182916_5a56.md
@@ -1,0 +1,105 @@
+---
+id: checkpoint_5a56c20e
+timestamp: 2026-06-03T18:29:16.665Z
+tags:
+  - julie-rescue
+  - adr-0006
+  - test-support
+  - cycle-break
+  - julie-core
+git:
+  branch: julie-rescue
+  commit: 0b64cce0
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - Cargo.lock
+    - crates/julie-core/Cargo.toml
+    - crates/julie-core/src/lib.rs
+    - crates/julie-core/src/test_support/cleanup.rs
+    - crates/julie-core/src/test_support/db/mod.rs
+    - crates/julie-core/src/test_support/db/rows.rs
+    - crates/julie-core/src/test_support/db/tests.rs
+    - crates/julie-core/src/test_support/mod.rs
+    - crates/julie-core/src/test_support/tempdir.rs
+    - crates/julie-core/src/tests/database.rs
+    - crates/julie-core/src/tests/database/concurrency_wal.rs
+    - crates/julie-core/src/tests/database_row_mapping.rs
+    - crates/julie-core/src/tests/helpers.rs
+    - crates/julie-core/src/tests/mod.rs
+    - crates/julie-core/src/tests/vector_storage.rs
+    - crates/julie-test-support/Cargo.toml
+    - crates/julie-test-support/src/cleanup.rs
+    - crates/julie-test-support/src/db/mod.rs
+    - crates/julie-test-support/src/db/rows.rs
+    - crates/julie-test-support/src/db/tests.rs
+summary: "ADR-0006 implemented: test_support feature-gated module in julie-core,
+  cycle broken"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo nextest run -p julie-core: 118/118 pass"
+  - "cargo tree -p julie-core | grep julie-test-support: EMPTY (cycle broken)"
+  - "cargo tree -p julie-test-support | grep julie-core: shows julie-core
+    (one-way)"
+  - "grep pub fn file_info_builder: single source in test_support/db/rows.rs"
+  - "cargo build: green"
+  - editing suite 103/103 pass (re-export chain works)
+next: "Commit as `refactor(core): host test_support in julie-core behind
+  feature, break dev-dep cycle (ADR-0006)`"
+---
+
+## ADR-0006 implemented: test_support feature-gated module in julie-core, cycle broken
+
+### WHAT
+Implemented the full ADR-0006 design — moved test helper bodies into `julie_core::test_support`, made `julie-test-support` a thin re-export, broke the dev-dep cycle.
+
+**Files created in `crates/julie-core/src/test_support/`:**
+- `mod.rs` — mirrors julie-test-support public surface + `open_test_connection` inline
+- `db/mod.rs` — re-exports from rows
+- `db/rows.rs` — all builders (file_info_builder, symbol_builder, etc.) with `crate::database::*`
+- `db/tests.rs` — builder tests (from julie-test-support/src/db/tests.rs), rewritten to `crate::database::SymbolDatabase`
+- `tempdir.rs` — unique_temp_dir (exact copy)
+- `cleanup.rs` — atomic_cleanup_julie_dir (exact copy)
+
+**`crates/julie-core/src/lib.rs`:** added `#[cfg(any(test, feature = "test-support"))] pub mod test_support;`
+
+**`crates/julie-core/Cargo.toml`:**
+- Added `[features]` with `test-support = ["dep:tempfile"]`
+- Added `tempfile = { version = "3.13", optional = true }` to `[dependencies]`
+- Removed `julie-test-support = { path = "../julie-test-support" }` from `[dev-dependencies]` (cycle break)
+
+**`crates/julie-test-support`:**
+- `Cargo.toml`: reduced to only `julie-core = { path = "../julie-core", features = ["test-support"] }`
+- `src/lib.rs`: `pub use julie_core::test_support::*;` thin re-export
+- Deleted: `src/db/`, `src/tempdir.rs`, `src/cleanup.rs`
+
+**`crates/julie-core/src/tests/` import updates:**
+- Deleted `helpers.rs` (local duplicate — superseded by `crate::test_support`)
+- Removed `mod helpers;` from `mod.rs`
+- `database.rs`, `database_row_mapping.rs`, `vector_storage.rs`: `super::helpers::*` + `julie_test_support::*` → `crate::test_support::*`
+- `database/concurrency_wal.rs`: 3x `julie_test_support::open_test_connection` → `crate::test_support::open_test_connection`
+
+### WHY
+ADR-0006 decision: hosting helpers in the leaf crate eliminates the two-rlib type mismatch that prevented julie-core's own tests from using shared builders for julie-core-owned types (SymbolDatabase, FileInfo).
+
+### HOW
+- Feature gate `#[cfg(any(test, feature = "test-support"))]` makes helpers available in test builds and to downstream consumers via feature flag
+- `resolver = "2"` keeps the optional tempfile dep out of production builds
+- `julie-test-support` thin re-export preserves all existing import paths unchanged
+
+### IMPACT
+- `cargo nextest run -p julie-core`: **118/118 pass** (was 107 — gained 11 builder/tempdir tests)
+- `cargo tree -p julie-core | grep julie-test-support`: EMPTY (cycle broken)
+- `cargo tree -p julie-test-support | grep julie-core`: shows one-way dep
+- `grep "pub fn file_info_builder" src/ crates/`: single source in `test_support/db/rows.rs`
+- `cargo build`: green; `cargo build -p julie-core` (no features): green
+- Top-crate tests (editing suite 103/103) pass through the re-export chain
