+++ crates/julie-core/src/tests/database_row_mapping.rs
@@ -1,6 +1,5 @@
 use crate::database::SymbolDatabase;
-use super::helpers::{file_info_builder, store_file_info_if_missing};
-use julie_test_support::symbol_builder;
+use crate::test_support::{file_info_builder, store_file_info_if_missing, symbol_builder};
 use tempfile::TempDir;
 
 #[test]
