+++ crates/julie-core/src/tests/mod.rs
@@ -1,4 +1,3 @@
-mod helpers;
 mod database;
 mod database_init_race;
 mod database_lightweight_query;
